Fixed: CURLOPT_DNS_SERVERS in connection to Curl.duphandle in curl-helper.c
Given a handle with CURLOPT_DNS_SERVERS set, a duplicate would still point to the same memory location, as
the argument isn't set anew in duplicateConnection(..).
The original could be garbage collected and the space used otherwise, while the duplicate would still point to it,
leading to errors during Curl.perform on the duplicate.