Make it possible to add a name to a MIME part
For compatibility with existing code, this adds a constructor
`CURLMIME_DATA_WITH_NAME` to `curlMIMEPartData`.  In that variant, if
`name` is `Some str`, the library uses `curl_mime_name` to give the part
a name.  Same for `filename` but with `curl_mime_filename`.

This enables the use of the library for uploading to S3 while avoiding
the deprecated `curl_form*` functions.