Release lock on cleanup because it might call callbacks
Curl multi callbacks are registered with the
[`CURLMOPT_SOCKETFUNCTION`][1] option of [`curl_multi_setopt`][2]:

    curl_multi_setopt(multi->handle,
                      CURLMOPT_SOCKETFUNCTION,
                      curlm_sock_cb);

This `curlm_sock_cb` function is C code, that starts by acquiring the
OCaml [runtime lock][3] so that it can execute an OCaml callback, then
releases the runtime lock:

    static int curlm_sock_cb(CURL *e, curl_socket_t sock, int what,
                             void *cbp, void *sockp)
    {
      caml_acquire_runtime_system();

      /* ... */

      caml_callback2(Field(((ml_multi_handle*)cbp)->values,
                     curlmopt_socket_function), csock, v_what);

      /* ... */

      caml_release_runtime_system();
      return 0;
    }

It's good that these curl callback functions try to acquire the
runtime lock because they can be called from a context where the lock
isn't held.

During execution of `caml_curl_multi_cleanup`, the runtime lock is
held. The `curl_multi_cleanup` function might call callbacks which try
to acquire the runtime lock. As such, execution will block when the
callback try to acquire the runtime lock.

This patch fixes the issue by releasing the runtime lock before curl
callbacks are called, since they each try to acquire the lock.

[1]: https://curl.se/libcurl/c/CURLMOPT_SOCKETFUNCTION.html
[2]: https://curl.se/libcurl/c/curl_multi_setopt.html
[3]: https://ocaml.org/manual/5.3/intfc.html#ss:parallel-execution-long-running-c-code